componentizando a página de pesquisa
