+++ src/components/Search/index.tsx
@@ -1,0 +1,15 @@
+import YearAutocomplete from './YearsAutocomplete';
+import MonthsAutocomplete from './MonthsAutocomplete';
+import AgencyTypeSelect from './AgencyTypeSelect';
+import AgencyAutocomplete from './AgencyAutocomplete';
+import CategorySelect from './CategorySelect';
+
+const Search = {
+  YearsAutocomplete: YearAutocomplete,
+  MonthsAutocomplete: MonthsAutocomplete,
+  AgencyTypeSelect: AgencyTypeSelect,
+  AgencyAutocomplete: AgencyAutocomplete,
+  CategorySelect: CategorySelect,
+};
+
+export default Search;
